refactor trigger poller worker
