+++ crates/ironclaw_triggers/src/worker/report.rs
@@ -1,0 +1,93 @@
+use ironclaw_host_api::{TenantId, Timestamp};
+use ironclaw_turns::TurnRunId;
+
+use crate::TriggerId;
+
+use super::TrustedTriggerSubmitFailureReason;
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct TriggerPollerTickReport {
+    pub now: Timestamp,
+    pub active_records: usize,
+    pub due_records: usize,
+    pub results: Vec<TriggerPollerFireReport>,
+}
+
+impl TriggerPollerTickReport {
+    pub(super) fn new(now: Timestamp) -> Self {
+        Self {
+            now,
+            active_records: 0,
+            due_records: 0,
+            results: Vec::new(),
+        }
+    }
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct TriggerPollerFireReport {
+    pub tenant_id: TenantId,
+    pub trigger_id: TriggerId,
+    pub fire_slot: Timestamp,
+    pub outcome: TriggerPollerFireOutcome,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum TriggerPollerFireOutcome {
+    Submitted {
+        run_id: TurnRunId,
+    },
+    Replayed {
+        original_run_id: TurnRunId,
+    },
+    RetryableFailed {
+        reason: TriggerPollerFailureReason,
+    },
+    PermanentFailed {
+        reason: TriggerPollerFailureReason,
+    },
+    ClearedTerminalActive {
+        run_id: TurnRunId,
+    },
+    ActiveRunLookupFailed {
+        run_id: TurnRunId,
+        reason: TriggerPollerFailureReason,
+    },
+    SkippedAlreadyCleared {
+        run_id: TurnRunId,
+    },
+    SkippedAlreadyActive {
+        active_fire_slot: Timestamp,
+        active_run_ref: Option<TurnRunId>,
+    },
+    DueFireFailed {
+        reason: TriggerPollerFailureReason,
+    },
+    SkippedNotDue,
+    SkippedNotFound,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum TriggerPollerFailureReason {
+    Backend,
+    InvalidTriggerId,
+    InvalidFireIdentityComponent,
+    InvalidRecord,
+    InvalidPollerConfig,
+    InvalidSchedule,
+    InvalidMaterialization,
+    NotFound,
+    SourceNoFire,
+    TrustedSubmitRetryable,
+    TrustedSubmitPermanent,
+    ActiveRunLookup,
+}
+
+impl TriggerPollerFailureReason {
+    pub(super) fn from_trusted_submit_failure(reason: TrustedTriggerSubmitFailureReason) -> Self {
+        match reason {
+            TrustedTriggerSubmitFailureReason::Retryable => Self::TrustedSubmitRetryable,
+            TrustedTriggerSubmitFailureReason::Permanent => Self::TrustedSubmitPermanent,
+        }
+    }
+}
